requirements.txt
Flask==3.0.0
flask-cors==4.0.0
boto3==1.34.28
python-dotenv==1.0.0
Werkzeug==3.0.1
flask-limiter==3.5.0
flask-talisman==1.1.0
botocore>=1.34.28
PyMySQL==1.1.0
Flask-Login==0.6.3
bcrypt==4.1.2
google-api-python-client==2.108.0
requests==2.31.0
